prophet-agent/requirements.txt
pykrx>=1.0.45
opendartreader>=0.2.1
yfinance>=0.2.31
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
